+++ migrations/16_add_media_attributes_table_down.sql
@@ -1,0 +1,3 @@
+DROP INDEX idx_media_attributes_purpose;
+DROP INDEX idx_media_attributes;
+DROP TABLE media_attributes;