+++ types/media_attributes.go
@@ -1,0 +1,12 @@
+package types
+
+type MediaAttributes struct {
+	Origin  string
+	MediaId string
+	Purpose string
+}
+
+const PurposeNone = "none"
+const PurposePinned = "pinned"
+
+var AllPurposes = []string{PurposeNone, PurposePinned}